src/lambda/document_processor/requirements.txt
boto3>=1.26.0
psycopg2-binary>=2.9.5
pytesseract>=0.3.10
pdf2image>=1.16.3
PyPDF2>=3.0.1
Pillow>=9.4.0
pandas>=1.3.5
numpy>=1.22.0
tabula-py>=2.5.0
matplotlib>=3.5.0
opencv-python>=4.5.5
